+++ tests/bluetooth/tester/src/gatt.c
@@ -64,7 +64,7 @@ DECL|permissions|member|uint8_t permissions;
 DECL|permissions|member|uint8_t permissions;
 DECL|prep_data|member|uint8_t *prep_data;
 DECL|properties|member|uint8_t properties;
-DECL|read_cb|function|static uint8_t read_cb(struct bt_conn *conn, int err, struct bt_gatt_read_params *params, const void *data, uint16_t length)
+DECL|read_cb|function|static uint8_t read_cb(struct bt_conn *conn, uint8_t err, struct bt_gatt_read_params *params, const void *data, uint16_t length)
 DECL|read_destroy|function|static void read_destroy(struct bt_gatt_read_params *params)
 DECL|read_long|function|static void read_long(uint8_t *data, uint16_t len)
 DECL|read_multiple|function|static void read_multiple(uint8_t *data, uint16_t len)
